backend/requirements.txt
# Core Backend
fastapi
uvicorn

# Authentication & Security
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0

# Database
sqlalchemy

# Request Handling
pydantic
python-multipart
email-validator


# Frontend
streamlit